function `translate` returned to `Edges`
`width` and `height` fields are now private to ensure `BingImage` is
immutable. To get `width` and `height` there are methods of the same
name in `BinImage`. The `points_to_drawing_order` function has been
moved out of the `Edges` implementation, as the usage of the `translate`
argument has been moved to `image_edges`.
